Reorganize persistence modules
